% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_format.R
\name{da_format}
\alias{da_format}
\title{Format string to display data provided with the datalegreya font}
\usage{
da_format(.tbl, x, y, text, xlab = FALSE, ylab = FALSE)
}
\arguments{
\item{.tbl}{Dataframe containing data to display}

\item{x}{Name of x column of data}

\item{y}{Name of y column of data}

\item{xlab}{Boolean. Whether to display x axis. Default to FALSE.}

\item{ylab}{Boolean. Whether to display y axis. Default to FALSE.}
}
\description{
Format string to display data provided with the datalegreya font
}
\examples{
data <- data.frame(year = seq(2008, 2018, 1), val = sin(seq(0, pi, length.out = 11)))
t <- "datalegreya"
da_format(data, year, val, t)

}
