% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fig}
\alias{fig}
\title{Insert a datalegreya font easily into html outputs using R.}
\usage{
fig(object, symbol = NULL, ylabs = rep(NA, 2), xlabs = rep(NA, 2),
  start_value = 0, size = NULL, type = "dot", color = NULL,
  colour = color, bgcolor = NULL, bgcolour = bgcolor, sother = NULL)
}
\arguments{
\item{object}{A named integer vector with elements of 0, 1, 2, 3 and names corresponding to the characters of the text.}

\item{symbol}{A named integer vector where the name should signify which symbol (min, max, down or up) and the integer specifying above which character it should be placed.}

\item{ylabs}{A string vector of length 2 comprising of string of max length 5 for the bottom and top label of the y-axis. This is NOT working for some reason.}

\item{xlabs}{A string vector of length 2 comprising of string (of max length 5) to be placed on top of the main text. The first string is placed on the start of the main text and the second at the end. Leave as NA to have one or both to display no labels.}

\item{start_value}{An integer between 0 and 3 (inclusive) to specify the starting value of the graph.}

\item{size}{The size of the font.}

\item{type}{The type of datalegreya font, either thin, gradient or dot.}

\item{color, colour}{Colour to be given to the icon}

\item{bgcolor, bgcolour}{Colour to be given to the background}

\item{sother}{Character vector of other parameters directly added to the style classes.}
}
\description{
Insert a datalegreya font easily into html outputs using R.
}
\references{
\href{https://github.com/figs-lab/datalegreya}{Datalegreya}
}
